♻️ refactor(book/collection): move book collection components to new collection directory
- rename all "list" references to "collection" in component imports
- update imports for BookCollection and BookCollectionSkeleton in index.tsx and book-collection.tsx
