+++ apps/tk-capture-book/features/book/components/collection/book-collection-skeleton.tsx
